Fixed failing test
I'm a bit stomped by this test failure as it didn't show up in the CI of PR:
https://github.com/mauro3/SimpleTraits.jl/pull/34

Furthermore, the same issue existed as far back as
https://github.com/mauro3/SimpleTraits.jl/pull/25

I tried to see whether this was introduced in one of the minor
versions of Julia 0.5.  But running the test now locally fails for all
0.5.0, 0.5.1 and 0.5.2.

So, not sure what changed.